rename directories, configure css module values for theming
