--- app/config.js
@@ -1,10 +1,0 @@
-export const autocompleteOptions = {
-    types: ['establishment'],
-    bounds: {
-      south: 40.734634,
-      north: 40.752200,
-      west: -74.002601,
-      east: -73.981465
-    },
-    strictBounds: true
-  }
