+++ client/index.js
@@ -4,13 +4,15 @@ import { Provider } from 'react-redux'
 import { Router, browserHistory } from 'react-router'
 import store from './redux'
 import routes from './routes'
-
-
-require('./auth').setupInterceptors()
+import theme from './theme'
+import { ThemeProvider } from 'react-css-themr'
+require('./utilities/auth').setupInterceptors()
 
 render(
+  <ThemeProvider theme={theme}>
     <Provider store={store}>
       <Router history={browserHistory} routes={routes} />
-    </Provider>,
+    </Provider>
+  </ThemeProvider>,
   document.getElementById('app')
 )
