+++ client/react/single-place/components/Item.js
