+++ client/react/single-place/components/RatingStars.js
