+++ client/react/single-place/components/index.js
