+++ client/react/single-place/index.js
