+++ client/theme/css/button.css
@@ -1,0 +1,15 @@
+
+@value buttonPrimaryColor, buttonPrimaryBackgroundColor, buttonPrimaryDisabledBackgroundColor from "./variables.css";
+
+@value button-border-radius: 4px;
+@value button-squared-min-width: 12em;
+
+.primary {
+  float: right;
+  color: buttonPrimaryColor !important;
+  background-color: buttonPrimaryBackgroundColor !important;
+
+  &.flat {
+   background-color: buttonPrimaryDisabledBackgroundColor !important;
+ }
+}
