+++ client/theme/css/variables.css
@@ -1,0 +1,16 @@
+@value buttonPrimaryColor: #000;
+@value buttonPrimaryBackgroundColor: #fff;
+@value buttonPrimaryDisabledBackgroundColor: #ccc;
+
+/* global React-Toolbox overrides - THESE MUST BE NAMED EXACTLY LIKE THIS */
+@value color-divider: var(--palette-grey-200);
+@value color-background: var(--color-white);
+@value color-text: var(--palette-grey-900);
+@value color-text-secondary: var(--palette-grey-600);
+
+@value color-primary: var(--palette-indigo-500);
+@value color-primary-dark: var(--palette-indigo-700);
+@value color-accent: var(--palette-pink-a200);
+@value color-accent-dark: var(--palette-pink-700);
+@value color-primary-contrast: var(--color-dark-contrast);
+@value color-accent-contrast: var(--color-dark-contrast);
