+++ client/theme/index.js
@@ -1,0 +1,3 @@
+import RTButton from './css/button.css'
+
+export default { RTButton }
