+++ package.json
@@ -39,19 +39,18 @@
     "express": "^4.14.1",
     "express-session": "^1.15.1",
     "lodash": "^4.17.4",
+    "modules-values-extract": "^1.0.3",
     "morgan": "^1.7.0",
     "passport": "^0.3.2",
     "passport-google-oauth": "^1.0.0",
     "passport-local": "^1.0.0",
     "path": "^0.12.7",
     "postcss-cssnext": "^2.10.0",
-    "postcss-each": "^0.9.3",
-    "postcss-import": "^9.1.0",
     "postcss-loader": "^1.3.3",
-    "postcss-mixins": "^5.4.1",
     "promise-mysql": "^3.0.0",
     "react": "^15.4.2",
     "react-addons-css-transition-group": "^15.4.2",
+    "react-css-themr": "^2.0.0",
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.2",
     "react-router": "^3.0.3",
