+++ postcss.config.js
@@ -1,11 +1,29 @@
+const fs = require('fs')
+const path = require('path')
+const extractValues = require('modules-values-extract')
 
-module.exports = {
-  plugins: {
-    'postcss-import': {
-      root: __dirname,
-    },
-    'postcss-mixins': {},
-    'postcss-each': {},
-    'postcss-cssnext': {}
-  },
-}
+
+const config = (variables) => ({
+    plugins: {
+      'postcss-cssnext': {
+        features: {
+            customProperties: { variables }
+          }
+      },
+      'postcss-modules-values': {}
+    }
+  })
+
+const cssPath = path.join(__dirname, 'client', 'theme', 'css')
+const cssFiles = fs
+                  .readdirSync(cssPath)
+                  .filter(file => file.match(/\.css/i))
+                  .map(file => path.resolve(path.join(cssPath, file)))
+
+
+module.exports = extractValues({ files: cssFiles }).then((cssVars) => config(
+  Object
+    .keys(cssVars)
+    .filter(key => key.match(/-/))
+    .reduce((acc, key) => Object.assign({}, acc, {[key]: cssVars[key]}), {})
+))
