+++ public/style.css
@@ -1,5 +1,6 @@
 .tagline {
   text-align: center;
+  color: #000;
 }
 
 .google-sign-in {
